# [proto] Mark the fields in Prometheus remote read/write TimeSeries as nonnullable