--- a/src/query/generated/proto/prompb/types.proto
+++ b/src/query/generated/proto/prompb/types.proto
@@ -12,8 +12,8 @@ message Sample {
 }
 
 message TimeSeries {
-  repeated Label labels   = 1;
-  repeated Sample samples = 2;
+  repeated Label labels   = 1 [(gogoproto.nullable) = false];;
+  repeated Sample samples = 2 [(gogoproto.nullable) = false];;
 }
 
 message Label {