--- a/src/query/storage/converter.go
+++ b/src/query/storage/converter.go
@@ -68,7 +68,7 @@ var (
 
 // PromLabelsToM3Tags converts Prometheus labels to M3 tags
 func PromLabelsToM3Tags(
-	labels []*prompb.Label,
+	labels []prompb.Label,
 	tagOptions models.TagOptions,
 ) models.Tags {
 	tags := models.NewTags(len(labels), tagOptions)
@@ -93,7 +93,7 @@ func PromLabelsToM3Tags(
 }
 
 // PromSamplesToM3Datapoints converts Prometheus samples to M3 datapoints
-func PromSamplesToM3Datapoints(samples []*prompb.Sample) ts.Datapoints {
+func PromSamplesToM3Datapoints(samples []prompb.Sample) ts.Datapoints {
 	datapoints := make(ts.Datapoints, 0, len(samples))
 	for _, sample := range samples {
 		timestamp := PromTimestampToTime(sample.Timestamp)
@@ -216,7 +216,7 @@ func (t sortableLabels) Less(i, j int) bool {
 }
 
 // TagsToPromLabels converts tags to prometheus labels.
-func TagsToPromLabels(tags models.Tags) []*prompb.Label {
+func TagsToPromLabels(tags models.Tags) []prompb.Label {
 	// Perform bulk allocation upfront then convert to pointers afterwards
 	// to reduce total number of allocations. See BenchmarkFetchResultToPromResult
 	// if modifying.
@@ -240,16 +240,12 @@ func TagsToPromLabels(tags models.Tags) []*prompb.Label {
 	// Sort here since name and label may be added in a different order in tags
 	// if default metric name or bucket names are overridden.
 	sort.Sort(sortableLabels(labels))
-	labelsPointers := make([]*prompb.Label, 0, l)
-	for i := range labels {
-		labelsPointers = append(labelsPointers, &labels[i])
-	}
 
-	return labelsPointers
+	return labels
 }
 
 // SeriesToPromSamples series datapoints to prometheus samples.SeriesToPromSamples.
-func SeriesToPromSamples(series *ts.Series) []*prompb.Sample {
+func SeriesToPromSamples(series *ts.Series) []prompb.Sample {
 	var (
 		seriesLen  = series.Len()
 		values     = series.Values()
@@ -266,12 +262,7 @@ func SeriesToPromSamples(series *ts.Series) []*prompb.Sample {
 		})
 	}
 
-	samplesPointers := make([]*prompb.Sample, 0, len(samples))
-	for i := range samples {
-		samplesPointers = append(samplesPointers, &samples[i])
-	}
-
-	return samplesPointers
+	return samples
 }
 
 func iteratorToTsSeries(