**What this PR does / why we need it**:

This PR adds the gogoproto `nonnullable` annotation to the fields of `TimeSeries` in the Prometheus remote read/write Protobuf type definitions. The [upstream Prometheus type definition](https://github.com/prometheus/prometheus/blob/master/prompb/types.proto) already sets this annotation so this PR brings our internal types closer in line with upstream. In addition, by changing the type of the `TimeSeries` fields from pointers to structs to just the structs themselves we are able to reduce the number of allocations and improve the speed of converting to and from M3's internal metrics types to the Prometheus remote read/write metrics types:

```
$ benchcmp ~/old.txt ~/new.txt
benchmark                               old ns/op     new ns/op     delta
BenchmarkFetchResultToPromResult-12     8650263       5531520       -36.05%

benchmark                               old allocs     new allocs     delta
BenchmarkFetchResultToPromResult-12     5652           3423           -39.44%

benchmark                               old bytes     new bytes     delta
BenchmarkFetchResultToPromResult-12     25519493      17143505      -32.82%
```

**Special notes for your reviewer**:

**Does this PR introduce a user-facing and/or backwards incompatible change?**:

```release-note
NONE
```

**Does this PR require updating code package or user-facing documentation?**:

```documentation-note
NONE
```